4112-51997-Math2-result1575168016472

+++ src/main/java/org/apache/commons/math3/analysis/FunctionUtils.java
@@ -436,7 +436,7 @@ public class FunctionUtils {
         }
 
         final double[] s = new double[n];
-        final double h = (max - min) / n;
+        final double h = (double)((max - min) / n);
         for (int i = 0; i < n; i++) {
             s[i] = f.value(min + i * h);
         }
